chore: Added Row component for structure chore: Improved AnchorButton refactor: Renamed LinkButton to AnchorButton refactor: Moved HTMLComponents to src directory fix: Removed register text from profile page
