+++ src/Login.js
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 import {
-  Container, Flashes, Form, Input, Submit,
+  Container, Flashes, Form, Input, Row, Submit,
 } from './Components';
 
 export default function Login(props) {
@@ -17,35 +17,38 @@ export default function Login(props) {
   return (
     <Container>
       <h1>Log In</h1>
-      <Form token={csrfToken}>
-        <Input
-          type="text"
-          id="username"
-          placeholder="Username"
-          errors={formErrors}
-          required
-        />
 
-        <Input
-          type="password"
-          id="password"
-          placeholder="Password"
-          errors={formErrors}
-          required
-        />
+      <Row>
+        <Form token={csrfToken}>
+          <Input
+            type="text"
+            id="username"
+            placeholder="Username"
+            errors={formErrors}
+            required
+          />
 
-        <Submit id="submit" value="Login" />
-      </Form>
+          <Input
+            type="password"
+            id="password"
+            placeholder="Password"
+            errors={formErrors}
+            required
+          />
 
-      <br />
+          <Submit id="submit" value="Login" />
+        </Form>
+      </Row>
 
       <Flashes flashes={flashes} />
 
-      Don&apos;t have an account? Register
-      {' '}
-      {/* TODO: replace this with a React Router Link */}
-      <a href="/registration">here</a>
-      .
+      <Row>
+        Don&apos;t have an account? Register
+        {' '}
+        {/* TODO: replace this with a React Router Link */}
+        <a href="/registration">here</a>
+        .
+      </Row>
     </Container>
   );
 }
