+++ src/Registration.js
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 import {
-  Container, Flashes, Form, Input, Submit,
+  Container, Flashes, Form, Input, Row, Submit,
 } from './Components';
 
 export default function Registration(props) {
@@ -17,39 +17,42 @@ export default function Registration(props) {
   return (
     <Container>
       <h1>Sign Up</h1>
-      <Form token={csrfToken}>
-        <Input
-          type="text"
-          id="username"
-          placeholder="Username"
-          errors={formErrors}
-          minlength="4"
-          maxlength="20"
-          required
-        />
 
-        <Input
-          type="password"
-          id="password"
-          placeholder="Password"
-          errors={formErrors}
-          minlength="4"
-          maxlength="20"
-          required
-        />
+      <Row>
+        <Form token={csrfToken}>
+          <Input
+            type="text"
+            id="username"
+            placeholder="Username"
+            errors={formErrors}
+            minlength="4"
+            maxlength="20"
+            required
+          />
 
-        <Submit id="submit" value="Login" />
-      </Form>
+          <Input
+            type="password"
+            id="password"
+            placeholder="Password"
+            errors={formErrors}
+            minlength="4"
+            maxlength="20"
+            required
+          />
 
-      <br />
+          <Submit id="submit" value="Login" />
+        </Form>
+      </Row>
 
       <Flashes flashes={flashes} />
 
-      Already have an account? Log in
-      {' '}
-      {/* TODO: replace this with a React Router Link */}
-      <a href="/login">here</a>
-      .
+      <Row>
+        Already have an account? Log in
+        {' '}
+        {/* TODO: replace this with a React Router Link */}
+        <a href="/login">here</a>
+        .
+      </Row>
     </Container>
   );
 }
